fix: address review findings for storage upload duplication
- Extract shared uploadStorageServerSide() to eliminate ~140 lines of
  duplication between skill-upload.ts and instruction-upload.ts
- Remove defensive try-catch in deleteSkillServerSide S3 cleanup,
  letting errors propagate to the caller per project conventions
